fix endTurn logic
End turn was misbehaving because of gameState.turn not updating before the function tried to modify state. Using prev.turn inside setGameState results in correct reference for updating the turn when the function is called.

For separation of concerns, moved the call to generateHand from inside the endTurn function to a separate useEffect hook which depends on gameState.turn