modify starting match logic
Design choice: randomly assign who gets the first turn at the start of the match, and generate cards only for that player. Going first is an advantage, but the first hand of cards has no previous game state to guide card generation. Going second is hopefully balanced out by receiving a hand of cards that is adapted to actions taken in the first turn.